Formatted

+++ src/@crema/core/AppsDeleteIcon/index.js
@@ -1,32 +1,24 @@
-import React, {useState} from 'react';
-import {AiOutlineDelete} from 'react-icons/ai';
+import React, { useState } from 'react';
+import { AiOutlineDelete } from 'react-icons/ai';
 import IntlMessages from '../../utility/IntlMessages';
 import PropTypes from 'prop-types';
 import ConfirmationModal from '../AppConfirmationModal';
 import AppIconButton from '../AppIconButton';
 
-const AppsDeleteIcon = ({deleteAction}) => {
+const AppsDeleteIcon = ({ deleteAction }) => {
   const [isDeleteModalOpen, setDeleteModalOpen] = useState(false);
 
   return (
     <>
       <AppIconButton
-        title='Trash'
+        title="Trash"
         icon={<AiOutlineDelete />}
         onClick={() => {
           console.log('mai hu');
           setDeleteModalOpen(true);
         }}
       />
-      {isDeleteModalOpen ? (
-        <ConfirmationModal
-          open={isDeleteModalOpen}
-          paragraph={'Are you sure you want to delete this?'}
-          onDeny={setDeleteModalOpen}
-          onConfirm={deleteAction}
-          modalTitle={<IntlMessages id='common.deleteItem' />}
-        />
-      ) : null}
+      {isDeleteModalOpen ? <ConfirmationModal open={isDeleteModalOpen} paragraph={'Are you sure you want to delete this?'} onDeny={setDeleteModalOpen} onConfirm={deleteAction} modalTitle={<IntlMessages id="common.deleteItem" />} /> : null}
     </>
   );
 };
